Move JOB_STATUS constant to break import cycle
